Moved types to vite-env.d.ts and upadated Card.tsx ProductList.tsx
