+++ client/src/pages/ProductList.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import Card from "../components/products/Card.tsx";
-import type { ProductType } from "../types/ClientType.ts";
+import type { ProductType } from "../types/vite-env.d.ts";
 
 function Products() {
   const [products, setProducts] = useState<ProductType[] | null>(null);
