--- client/src/types/ClientType.ts
@@ -1,12 +1,0 @@
-export type ProductType = {
-  id: number;
-  name: string;
-  description: string;
-  price: number;
-  category_id: number;
-  images: string;
-};
-
-export type CardProps = {
-  products: ProductType[] | null;
-};
